backend/Database/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
boto3==1.29.7
botocore==1.32.7
pyjwt==2.10.1
gunicorn==21.2.0
